requirements.txt
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
idna==3.7
netifaces==0.10.6
pillow==10.3.0
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
socketIO-client==0.7.2
urllib3==2.2.1
websocket-client==1.7.0
wsproto==1.2.0
